Revert "Added multimaps option for testing (#1510)"
This reverts commit 8085b158fc02f86352b664f417a0f74744deb64d.

+++ Sources/CyCity.cpp
@@ -594,7 +594,7 @@ void CyCity::changePopulation(int iChange)
 	m_pCity->changePopulation(iChange);
 }
 
-int64_t CyCity::getRealPopulation() const
+int CyCity::getRealPopulation() const
 {
 	return m_pCity->getRealPopulation();
 }
